% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hourly_rad.R
\name{hourly_rad}
\alias{hourly_rad}
\title{hourly_rad}
\usage{
hourly_rad(hour, df1)
}
\arguments{
\item{hour}{Desired hour for which irradiation values are returned}

\item{df1}{Data frame filtered to the desired Month-Day combination}
}
\description{
Returns values of the Month/Day combinations which correspond to the desired hour
}
\author{
Philip Kezich
}

