% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postcode_pv_output.R
\name{postcode_pv_output}
\alias{postcode_pv_output}
\title{postcode_pv_output}
\usage{
postcode_pv_output(Postcode, MonthDay, data.folder, InstalledPV, temp_vector)
}
\arguments{
\item{Postcode}{Desired Postcode}

\item{MonthDay}{Desired Month/Day combination}

\item{data.folder}{Folder containing the processed Irradiation data}

\item{InstalledPV}{Data Frame containing the Installed PV values sourced from the
clean energy regulator}

\item{temp_vector}{Vector of hourly temperature values for our selected day}
}
\description{
This function returns the mean and upper and lower bounds of the total electricity output
for a desired postcode on our selected Month/Day combination. This function relies on the function
hourly_rad and uses the temperature output of the temperature_interpolation function as an input
}

