% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temp_after_9am.R
\name{temp_after_9am}
\alias{temp_after_9am}
\title{temp_after_9am}
\usage{
temp_after_9am(hour, max, min)
}
\arguments{
\item{hour}{Desired hour of the day (From 10am to midnight)}

\item{max}{Maximum daily temperature}

\item{min}{Minimum daily temperature}
}
\value{
The temperature value corresponding to our selected hour
}
\description{
The function temp_till_9am is used to temperatures for times after 9am.
}
\details{

}
\author{
Philip Kezich
}

