% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temp_till_9am.R
\name{temp_till_9am}
\alias{temp_till_9am}
\title{temp_till_9am}
\usage{
temp_till_9am(hour, max, min)
}
\arguments{
\item{hour}{Desired hour of the day (up to and including 9am)}

\item{max}{Maximum daily temperature}

\item{min}{Minimum daily temperature}
}
\value{
The temperature value corresponding to our selected hour
}
\description{

}
\author{
Philip Kezich
}

