% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_postcode_pv.R
\name{calculate_postcode_pv}
\alias{calculate_postcode_pv}
\title{calculate_postcode_pv}
\arguments{
\item{pv_data}{Data folder for installed PV and geographic data}

\item{irradiation_data}{Data folder for processed irradiation data}

\item{Postcode}{Desired Postcode String}

\item{MonthDay}{Desired Month-Day String}
}
\value{
The total PV production for the mean and upper and lower bounds of daily solar irradiation, including a graph of the results
}
\description{
This function produces an estimate for the mean and upper and lower bounds of the 90th
percentile for PV output in a specified postcode usign the solaR model of a small scale PV system
}
\details{

}
\examples{
calculate_postcode_pv('C:/SolarPVUWA/data/Processed Data/','6004','06-21')
}
\author{
Philip Kezich
}

