% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_postcode_pv.R
\name{calculate_postcode_pv}
\alias{calculate_postcode_pv}
\title{This script calculated the mean and standard deviation for a month-day combination using the processed values
created by the massSolarCruncher script by Grant Coble-Neale. The script then  calculated the outputs of PV
installations in the specified postcode for cases corresponding to the mean and upper and lower bounds of the 90th
percentile for irradiation}
\usage{
calculate_postcode_pv(pv_data, irradiation_data, Postcode, MonthDay)
}
\arguments{
\item{pv_data}{Data folder for installed PV and geographic data}

\item{irradiation_data}{Data folder for processed irradiation data}
}
\description{
This script calculated the mean and standard deviation for a month-day combination using the processed values
created by the massSolarCruncher script by Grant Coble-Neale. The script then  calculated the outputs of PV
installations in the specified postcode for cases corresponding to the mean and upper and lower bounds of the 90th
percentile for irradiation
}
\author{
Philip Kezich
}

