% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temperature_interpolation.R
\name{temperature_interpolation}
\alias{temperature_interpolation}
\title{temperature_interpolation}
\usage{
temperature_interpolation(Postcode, MonthDay, pcode_lats_and_longs, bom_temps,
  bom_locs)
}
\description{
temperature_interpolation
}
\author{
Philip Kezich
}

